android prebuilds: Implement npm packaging
Not tested on darwin
Not tested with create-socket-app

Implement _build_remove_path to get past gradle folder handle locks on
Win32

+++ bin/publish-npm-modules.sh
@@ -9,6 +9,13 @@ declare dry_run=0
 declare only_platforms=0
 declare only_top_level=0
 
+function _publish () {
+  if (( !dry_run )); then
+    npm publish "${args[@]}" || exit $?
+  else
+    npm pack "${args[@]}" || exit $?
+  fi
+}
 
 if [[ "$platform" = "linux" ]]; then
   if [ -n "$WSL_DISTRO_NAME" ] || uname -r | grep 'Microsoft'; then
@@ -40,16 +47,23 @@ while (( $# > 0 )); do
     continue
   fi
 
+  if [[ "$arg" = "--no-rebuild" ]]; then
+    no_rebuild=1
+    continue
+  fi
+
   args+=("$arg")
 done
 
-rm -rf "$SOCKET_HOME"
-mkdir -p "$SOCKET_HOME"
+if (( !no_rebuild )); then
+  rm -rf "$SOCKET_HOME"
+  mkdir -p "$SOCKET_HOME"
+fi
 
 export SOCKET_HOME
 export PREFIX
 
-if (( !only_top_level )); then
+if (( !only_top_level ))  && (( !no_rebuild )) ; then
   "$root/bin/install.sh" || exit $?
 fi
 
@@ -87,6 +101,11 @@ if (( !only_top_level )); then
     cp -rf "$SOCKET_HOME/src"/* "$SOCKET_HOME/packages/$package/src"
     cp -rf "$SOCKET_HOME/include"/* "$SOCKET_HOME/packages/$package/include"
 
+    # don't copy debug files, too large
+    rm -rf $SOCKET_HOME/lib/*-android/objs-debug 
+    cp -rf $SOCKET_HOME/lib/*-android "$SOCKET_HOME/packages/$package/lib"
+    # cp -f $SOCKET_HOME/lib/*-android/*.a "$SOCKET_HOME/packages/$package/lib"
+
     cp -rf "$SOCKET_HOME/lib/"$arch-* "$SOCKET_HOME/packages/$package/lib"
     cp -rf "$SOCKET_HOME/objects/"$arch-* "$SOCKET_HOME/packages/$package/objects"
 
@@ -105,12 +124,7 @@ if (( !only_top_level )); then
     cd "$SOCKET_HOME/packages/$package" || exit $?
     echo "# in directory: '$SOCKET_HOME/packages/$package'"
 
-    if (( !dry_run )) ; then
-      npm publish "${args[@]}" || exit $?
-    else
-      # echo "# npm publish ${args[@]}"
-      npm pack "${args[@]}" || exit $?
-    fi
+    _publish
   done
 fi
 
@@ -118,10 +132,5 @@ if (( !only_platforms || only_top_level )); then
   cd "$SOCKET_HOME/packages/@socketsupply/socket" || exit $?
   echo "# in directory: '$SOCKET_HOME/packages/@socketsupply/socket'"
 
-  if (( !dry_run )); then
-    npm publish "${args[@]}" || exit $?
-  else
-    # echo "# npm publish ${args[@]}"
-    npm pack "${args[@]}" || exit $?
-  fi
+  _publish
 fi
